zng_ext_l10n

Struct LangMap

pub struct LangMap<V> { /* private fields */ }

Represents a map of Lang keys that can be partially matched.

Implementations

impl<V> LangMap<V>

pub fn new() -> Self

New empty default.

pub fn with_capacity(capacity: usize) -> Self

New empty with pre-allocated capacity.

pub fn best_match(&self, lang: &Lang) -> Option<&Lang>

Returns the best match to lang currently in the map.

pub fn get(&self, lang: &Lang) -> Option<&V>

Returns the best match for lang.

pub fn get_exact(&self, lang: &Lang) -> Option<&V>

Returns the exact match for lang.

pub fn get_mut(&mut self, lang: &Lang) -> Option<&mut V>

Returns the best match for lang.

pub fn get_exact_mut(&mut self, lang: &Lang) -> Option<&mut V>

Returns the exact match for lang.

pub fn get_exact_or_insert( &mut self, lang: Lang, new: impl FnOnce() -> V, ) -> &mut V

Returns the current value or insert new and return a reference to it.

pub fn insert(&mut self, lang: Lang, value: V) -> Option<V>

Insert the value with the exact match of lang.

Returns the previous exact match.

pub fn remove(&mut self, lang: &Lang) -> Option<V>

Remove the exact match of lang.

pub fn remove_all(&mut self, lang: &Lang) -> usize

Remove all exact and partial matches of lang.

Returns a count of items removed.

pub fn pop(&mut self) -> Option<(Lang, V)>

Remove the last inserted lang.

pub fn is_empty(&self) -> bool

Returns if the map is empty.

pub fn len(&self) -> usize

Returns the number of languages in the map.

pub fn clear(&mut self)

Remove all entries.

pub fn keys(&self) -> impl ExactSizeIterator<Item = &Lang>

Iterator over lang keys.

pub fn values(&self) -> impl ExactSizeIterator<Item = &V>

Iterator over values.

pub fn values_mut(&mut self) -> impl ExactSizeIterator<Item = &mut V>

Iterator over values.

pub fn into_values(self) -> impl ExactSizeIterator<Item = V>

Into iterator of values.

pub fn iter(&self) -> impl ExactSizeIterator<Item = (&Lang, &V)>

Iterate over key-value pairs.

pub fn iter_mut(&mut self) -> impl ExactSizeIterator<Item = (&Lang, &mut V)>

Iterate over key-value pairs with mutable values.

impl<V> LangMap<HashMap<LangFilePath, V>>

pub fn get_file(&self, lang: &Lang, file: &LangFilePath) -> Option<&V>

Returns the match for lang and file.

pub fn best_file_match( &self, lang: &Lang, file: &LangFilePath, ) -> Option<&LangFilePath>

Returns the best match to lang and file currently in the map.