Crate zng_ext_l10n

Localization service, l10n! and helpers.

Crate

This crate is part of the zng project.

Macros

l10n

Gets a variable that localizes and formats the text in a widget context.

lang

Compile-time validated Lang value.

Structs

FluentParserErrors

Errors found parsing a fluent resource file.

L10N

Localization service.

L10nDir

Represents localization resources synchronized from files in a directory.

L10nManager

Application extension that provides localization.

L10nMessageBuilder

Localized message variable builder.

Lang

Identifies the language, region and script of text.

LangFilePath

Localization resource file path in the localization directory.

LangMap

Represents a map of Lang keys that can be partially matched.

LangResource

Handle to a localization resource.

LangResources

Handle to multiple localization resources.

Langs

List of languages, in priority order.

NilL10nSource

Localization source that is never available.

SwapL10nSource

Represents localization source that can swap the actual source without disconnecting variables taken on resources.

Enums

L10nArgument

Represents an argument value for a localization message.

LangResourceStatus

Status of a localization resource.

Statics

LANG_VAR

Language of text in a widget context.

Traits

L10nSource

Represents a localization data source.